Added create_menubar
Added a menu bar to the top of the root with 'file' and 'edit' menu options.

File contains "Load Wardrobe" which will allow the user to choose the folder where all the clothes are saved.
It also contains "Exit" which exits the program

Edit contains "Add New Top" and "Add New Bottom" which allows the user to upload a new image to the wardrobe and choose which weather and occasion categories it falls under. The image will then automatically be added to the appropriate folder with the appropriate title.

The functions when these menus are pressed are still to be created