Save the file path
Made updates to __init__, load_wardrobe and update_wardrobe so that the program saves the wardrobe file path you last used. So when you open the program again it automatically loads the last wardrobe you had open.

It only asks you to choose a folder when you first open the program or when you want to load a new wardrobe